Changed ntracers to 10
The diagnostic value ntracers was set to 11 and was used in a loop
from 0 to (ntracers - 1) which wrote to a single-digit string. This
was fine with ifort but could upset some compilers. To avoid this,
we now just set ntracers to 10, but a better long-term solution
(used by GMAO) is just to use a longer string.